refactor(suite): remove MetadataLabeling component
