+++ packages/components/src/index.ts
@@ -13,7 +13,6 @@ export { Checkbox, type CheckboxProps } from './components/form/Checkbox/Checkbo
 export * from './components/animations/LottieAnimation';
 export { recolorLottieAnimation } from './components/animations/recolorLottieAnimation';
 export * from './components/Flag/Flag';
-export * from './components/AutoScalingInput/AutoScalingInput';
 export { Badge, type BadgeProps, type BadgeSize, type BadgeIntent } from './components/Badge/Badge';
 export * from './components/buttons/ButtonGroup/ButtonGroup';
 export { Button, type ButtonProps } from './components/buttons/Button/Button';
