+++ packages/suite/src/components/suite/labeling/Labeling/SuiteSyncInteractionsTooltip.tsx
