+++ packages/suite/src/components/suite/labeling/Labeling/selectIsLabelActionEnabled.test.ts
