+++ packages/suite/src/components/suite/labeling/index.tsx
@@ -3,5 +3,4 @@ export { AccountLabeling } from './AccountLabeling';
 export { AddressLabeling } from './AddressLabeling';
 export { WalletLabeling, useGetWalletLabel } from './WalletLabeling';
 // "User defined labeling"
-export { MetadataLabeling } from './MetadataLabeling/MetadataLabeling';
-export { Labeling } from './MetadataLabeling/MetadataLabeling';
+export { Labeling } from './Labeling/Labeling';
